requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-crispy-forms==1.12.0
django-filter==2.4.0
django-json-widget==1.1.1
django-sms==0.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
easy-thumbnails==2.7.1
future==0.18.2
httplib2==0.19.1
idna==3.2
ippanel==1.1.0
jdatetime==3.6.2
Markdown==3.3.4
messagebird==2.0.0
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
textmagic==2.0.3
urllib3==1.26.6
webcolors==1.11.1
